+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Tideland Go CouchDB Client
 
-## 2016-11-21
+## 2016-11-23
+
+- Edit *OpenPath()* to retrieve the nested configuration for
+  opening the database
+
+## 2016-11-22
 
 - Extended with bulk, design document, and view requests